feat: cursor left / right works as intended, up / down has problems with visual position vs internal position

+++ inc/Controller/Position.hpp
@@ -9,13 +9,11 @@
 #ifndef POSITION_HPP
 #define POSITION_HPP
 
-#include <stddef.h>
-
 struct Position {
-    size_t row;
-    size_t column;    
+    int row;
+    int column;    
 
-    Position(size_t r = 0, size_t c = 0): row{r}, column{c} {}
+    Position(int r = 0, int c = 0): row{r}, column{c} {}
 };
 
 #endif //POSITION_HPP
